공통 Modal Component width & height props 적용 & 모달 폴더구조 변경
