+++ src/components/Organisms/Modal/RegisterModal.tsx
@@ -15,9 +15,12 @@ import SelectInput from '../SelectInput';
 import { MdEdit } from 'react-icons/md';
 import CheckIcon from '@src/assets/check.svg';
 import { putUserInfo } from '@src/service/api';
+import { CreateModal } from '@src/components/Atoms/Modal';
 
 type RegisterModalProps = {
   onFinish: () => void;
+  onClose: () => void;
+  stopPropagation: (e: any) => void;
 };
 
 type Tech = {
@@ -29,7 +32,7 @@ type Tech = {
 //1. Check user nickname is already taken or not.
 //2. Upload Profile Image.
 //3. techList is hardcoded, it should go in DB i think.
-const RegisterModal = ({ onFinish }: RegisterModalProps) => {
+const RegisterModal = ({ onFinish, onClose, stopPropagation }: RegisterModalProps) => {
   const theme = useTheme();
   const [page, setPage] = useState<number>(0);
   const [user, setUser] = useRecoilState(userAtom);
@@ -225,35 +228,39 @@ const RegisterModal = ({ onFinish }: RegisterModalProps) => {
   }, [onCompleteSignUp, theme, user]);
 
   return (
-    <Container>
-      <Header>
-        {page !== 0 && (
-          <BackIcon
-            color={theme.colors.dark.a3}
-            size={24}
-            onClick={() => {
-              setPage((prev) => prev - 1);
-            }}
-          />
-        )}
-        <Label>
-          {[...Array(4)].map((_, index) => {
-            return (
-              <PageLabel key={index} active={page === index}>
-                {page === index ? index + 1 : ``}
-              </PageLabel>
-            );
-          })}
-        </Label>
-        <CloseModalButton onClick={onFinish} />
-      </Header>
-      <Divider />
+    <CreateModal onClick={onClose}>
+      <div onClick={stopPropagation}>
+        <Container>
+          <Header>
+            {page !== 0 && (
+              <BackIcon
+                color={theme.colors.dark.a3}
+                size={24}
+                onClick={() => {
+                  setPage((prev) => prev - 1);
+                }}
+              />
+            )}
+            <Label>
+              {[...Array(4)].map((_, index) => {
+                return (
+                  <PageLabel key={index} active={page === index}>
+                    {page === index ? index + 1 : ``}
+                  </PageLabel>
+                );
+              })}
+            </Label>
+            <CloseModalButton onClick={onFinish} />
+          </Header>
+          <Divider />
 
-      {page === 0 && NickNameContent()}
-      {page === 1 && TechContent()}
-      {page === 2 && ImageContent()}
-      {page === 3 && FinishContent()}
-    </Container>
+          {page === 0 && NickNameContent()}
+          {page === 1 && TechContent()}
+          {page === 2 && ImageContent()}
+          {page === 3 && FinishContent()}
+        </Container>
+      </div>
+    </CreateModal>
   );
 };
 
@@ -307,6 +314,7 @@ const FileInput = styled.input`
 `;
 
 const Header = styled.div`
+  position: relative;
   display: flex;
   flex-direction: row;
   width: 100%;
